push to roger

+++ firmware/src/firmware.c
@@ -148,7 +148,7 @@ int main(void) {
             print_str("b: X"); print_hex(b[h][w],4);
             print_str("a: X"); print_hex(a[h][w],4);
 
-            int pixel = (r[h][w] << 24) | (g[h][w] << 16) | (b[h][w] << 8) | a[h][w];
+            int pixel = ((unsigned int)r[h][w] << 24) | ((unsigned int)g[h][w] << 16) | ((unsigned int)b[h][w] << 8) | (unsigned int)a[h][w];
             print_str("Pizel: X");
             print_hex(pixel,8);
 
@@ -175,7 +175,7 @@ int main(void) {
 
 
 
-                value =  (r[h][w] << 24) | (g[h][w] << 16) | (b[h][w] << 8) | a[h][w];
+                value =  ((unsigned int)r[h][w] << 24) | ((unsigned int)g[h][w] << 16) | ((unsigned int)b[h][w] << 8) | (unsigned int)a[h][w];
                 if (running_array[index] == value) { //The pixel is in the running array
                     store_byte(&current, index, &image_chunk_index);
                 }
